voice-agent/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.24.0
sounddevice>=0.4.7
soundfile>=0.12.1
